requirements.txt
newspaper3k==0.2.8
aiohttp==3.8.5
kafka-python==2.0.2
pyspark==3.4.1
sentence-transformers==2.2.2
redis==4.6.0
numpy==1.24.3
fastapi
uvicorn==0.23.1